Handle "TRANSFERRING" state when waiting for FSx backup
When the FSx backup feature was launched backups remained in the
"CREATING" state (as per the "Lifecycle" attribute from the output of
the DescribeBackups API) until the attempted backup creation was
finished.

In order to address a bug, an additional "TRANSFERRING" state was added.
Thus, a successfully created backup is expected to pass through the
"CREATING" and then "TRANSFERRING" states.

Where before the FSx backup test assumed any state other than "CREATING"
meant backup creation had finished, this change enables the polling
functions to wait until its no longer in the "TRANSFERRING" state.

Signed-off-by: Tim Lane <tilne@amazon.com>